Fix two bugs found during overnight run + add smoke test
- _hash_split: handle single-boundary tuples (IndexError when OWT
  passed 1-element boundaries tuple for 2-way task/anchor split)
- LinearProbe.forward: cast activations to probe weight dtype to
  fix bf16/f32 mismatch during inner loop training
- Add smoke test config + script that skips Phase 0 baselines and
  mocks OWT download for fast end-to-end verification

+++ src/probes/probe.py
@@ -34,7 +34,7 @@ class LinearProbe(nn.Module):
         Returns:
             (batch,) logits (pre-sigmoid).
         """
-        return self.linear(h).squeeze(-1)
+        return self.linear(h.to(self.linear.weight.dtype)).squeeze(-1)
 
 
 class SklearnProbe:
